+++ test/examples/06-dynamic-add-remove.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { WidthProvider, Responsive } from "react-grid-layout";
+import { WidthProvider, Responsive } from "@sguisse/react-grid-layout";
 import _ from "lodash";
 const ResponsiveReactGridLayout = WidthProvider(Responsive);