+++ test/examples/test_demo.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import RGL, { WidthProvider } from 'react-grid-layout';
+import RGL, { WidthProvider } from '@sguisse/react-grid-layout';
 
 const ReactGridLayout = WidthProvider(RGL);